requirements.txt
fastapi
uvicorn
pydub
python-multipart
demucs
PyTorch